Assure we don't accidentally parse a valid-looking URL to `url` and cause long compute times.
The issue is that the `url` crate can be fooled by long URLs into parsing way-too-long hosts
inefficiently enough to allow DoS attacks by slowing it down.

Some predicate was implemented to detect such URLs before passing them to `url`, but unfortunately
a somewhat valid looking URL could be crafted to bypass that check and still run into the `url` parser,
which seemed to have no problem with the format of the URL.

Testing showed that this is a very special case, which is now handled as well. We hope that one day `url`
will handle this all by itself.