Initial implementation of topological commit walk
This commit introduces a new module `topo` in gix-traverse that contains
an iterator for walking commits topologically. It is very much based on
the existing Git implementation introduced in b45424181e[1] as well as
improvements done since then. Function names are kept as much as
possible. The mailing list[2] has a lot of interesting discussion about
how the algorithm works. In short it is a three stage process that aims
to reduce the amount of traversal needed before the first result can be
output.

Hopefully this makes it easier to use gitoxide to implement other tools
that need topological sorting, e.g. graphical log and blame.

For tests there is a separate type named TraversalAssertion. There is
probably grounds for making code more common in this area.

[1]: https://github.com/git/git/commit/b45424181e9e8b2284a48c6db7b8db635bbfccc8
[2]: https://public-inbox.org/git/pull.25.git.gitgitgadget@gmail.com/

Add tests for gix_traverse::topo

This unfortunately duplicates types found in the existing tests.