feat!: make it possible to trace incoming and outgoing packetlines.
Due to the way this is (and has to be) setup, unfortunately one
has to integrate that with two crates, instead of just one.

This changes touches multiple crates, most of which receive a single
boolean as last argument to indicate whether the tracing should
happen in the first place.