fix: V1 negotiation won't hang anymore (#1061)
The logic previously tried to estimate when a pack can be expected,
and when a NAK is the end of a block, or the beginning of a pack.

This can be known because a pack (with our settings) needs two things:

* the server thinks it's ready
* a `done` sent by the client

If the server doesn't think it's ready it will send NAK and be done.

So the logic should be, for a NAK to stop the read-loop, that the client
expects a pack, and the server is ready. If the client is not ready, or
the server isn't ready, keep NAK and consider them the end of a round,
hence break the loop.