add some tests to be clearer about the kind of command that is created (#1103)
Note that for a moment, there was the idea about using `shell-words`
split certain invocations ourselves to avoid having to use a shell.

This is great on windows, where a shell might not be available.
On linux though this also means that some shell-builtins might
not be not be found if this is used, which limits the approach
to windows, and even there it might reduce compatibility.

Thus, let's not do it.