remove `Clone` impl of `ItemSliceSend` and make it `Sync` instead
The `Clone` impl allows two `ItemSliceSend` to be backed by the same
slice. That makes the behavior of `get_mut()` harder to reason about
because we have to consider clones of the `ItemSliceSend` to determine
if it's safe. By making it `Sync`, we make it more obvious that
multiple threads referring to the same instance need to be considered.

I renamed the type to `ItemSliceSync` since it's now also `Sync`.

We could make the type also check the bounds and that an individual
element has not been borrowed more than once. Then we could make
`get_mut()` not `unsafe`.