more changelogs, this time for production crates
Really, one day, maybe this year, I should really, really figure out
how to prevent re-releasing stable crates all the time.

The answer is they must be independent of unstable crates, or they
there is a blanket rule that assume unstable APIs aren't exposed
so it can be a point-release. But if it's a point release, users
would automatically update to it and maybe that causes other
incompatibilities? Maybe not, as these unstable crates aren't exposed.