git-lfs might fail early; let's rely on these caches to be recreated, where possible
Maybe this also means that one should prefer to not use git-lfs at all
and keep data in the packs themselves, it still shouldn't be too much overall.

Butl... for now, let's try to keep it to make the experience.

Here we choose the quick way out by simply checking in the archives in question
and by disabling git-lfs for them.

That way, most of the scripts still run, but we don't have to square the circle
trying to get these scripts portable, or spend a lot of time figuring out why
some baselines don't match perfectly (on windows).