feat: add `config::Snapshot::trusted_program()`. (#1178)
That way it's possible to obtain an executable, program or script
from a key in the configuration that is in a trusted section of the
configuration.

This goes along with a new `command` feature that brings in the `command`
module at the top level to be able to execute such commands.