Avoid using #[cfg] on multiple individual function arguments
Attaching #[cfg] to individual arguments makes it look like the function
has five conditionally present arguments, and doesn't make it
immediately apparent that the first two are for the first argument and
the last three are for the second argument.

Split them into separate `let` statements for clarity.

In the process, factor out the common `.try_into().ok()?` from each.