assure the fuzzer can't construct expensive values
Previously the fuzzer would create a section with 15k keys
which all were `n`. The first `n` has a 136kb sized value
which triggers an expensive normalization.

All other `n` are empty, but previously caused accessing the
long value again which is the first `n` it finds.

Now we iterate values, and keep track of previous ones when
doing multi-value access, which seems like a good test to
have.