fix: allow parsing signatures with trailing numbers in offset (#1322)
That way this won't fail, but *just* silently degenerates information.

The idea is that during FSCK, objects should be decoded and then
re-encoded to see if they are still the same. If not, this means
some leniency kicked in.

Maybe for FSCK, there would also have to be a refactor so there
is a lenient and strict parsing mode. One problem at a time.