Add CI checks for copy-packetline
For the real purpose of this new check, which is to verify that
`just copy-packetline` has been run when needed, to ensure that
gix-packetline-blocking/src is not out of date, it should only be
necessary to run this check on one platform.

So this should probably be changed to `runs-on: ubuntu-latest`
instead of a matrix strategy soon. I'm running it on three platforms
initially to check the script (though some functionality of the
script is not exercised).

I've also tested the script locally on Ubuntu 22.04.4 LTS, and on
Windows 10 (in Git Bash).