Better explain why we cancel even on unstaged deletions
Only the secondary reason, which is the less important one, was
given before. Now both are given and the more important reason is
emphasized.

(If the entire subtree is deleted on disk, staged or not, then we
do proceed, since neither disadvantage applies. I think this is
clear from context, since this is checked first, and like the
other checks, it is commented.)