Add ambiguous user unit tests, and more for hostname
Not all of these tests can pass yet, since gix-transport does not
yet detect and refuse to proceed with leading-hypnen usernames.
Some pass; those that do not are, as expected:

- ambiguous_user_is_disallowed_explicit_ssh
- ambiguous_user_is_disallowed_implicit_ssh
- ambiguous_user_and_host_remain_disallowed_together_explicit_ssh
- ambiguous_user_and_host_remain_disallowed_together_implicit_ssh

This also adds AmbiguousUserName in one of the enums that will need
to have it, but nothing fails with this error yet; it is introduced
now only to facilitate writing unit tests that assert it.