Give `ArgumentSafety` traits; test `*_as_argument` methods
The three new `Url::*_as_argument` methods were only tested
indirectly before (and the path one not at all). This adds tests
for them so they have the same direct coverage as the corresponding
`*_argument_safe` methods.

The `Debug` and `PartialEq` traits added to `ArgumentSafety` are
primarily added to allow the new assertions to be written in a more
streamlined way (without matching).