Start on demo script making repo with NTFS stream
The repo this script makes attempts to check out entries traversing
the default `$INDEX_ALLOCATION` directory stream of the `.git`
directory, whose stream name is documented to be `$I30`.

However, although I am able to access directories under this naming
scheme through other applications, the repositories this script
currently creates do not appear to trigger the bug in gitoxide. The
next step is to try specifying the stream type explicitly.