fix!: assure that special device names on Windows aren't allowed.
Otherwise it's possible to read or write to devices when interacting
with references of the 'right' name.

This behaviour can be controlled with the new `prohibit_windows_device_names` flag,
which is adjustable on the `Store` instance as field, and which now has to be
passed during instantiation as part of the new `store::init::Options` struct.