fix: do not write empty packs and indices (#1404)
Failure can occour on Windows as it's likely such a pack or index is already opened
during negotiation. Then, when an empty fetch via `--depth 1` is repeated,
a temporary file would be renamed onto one an mmapped pack or index, causing
a failure.

Now packs or indices aren't written anymore if they are empty.