Conditionally compile gix-revision at_symbol fuzzed test
This allows compilation to succeed when building tests on 32-bit
Windows systems, and probably other 32-bit systems/builds.

This test function had been set as ignored in 32-bit builds with
the `cfg_attr` attribute. But it was still compiled, and the
compilation failed on systems where the literal 9223372036854775808
is too big for `usize`.

This was due to the implicit `#[deny(overflowing_literals)]`. While
that could be conditionally suppressed for 32-bit builds, the test
would not really be meaningful. So this leaves that diagnostic in
place, and turns the conditional ignore into conditional
compilation.