Remove execution-specific code in sourcing-only .sh files
Some of the .sh script files are only meant ever to be sourced and
not to be executed. These are already correctly not marked
executable, but some of them contained code that would only be
proper for scripts that are actually run, and not for scripts
that are only sourced:

- One contained a #! line, which this removes because it falsely
  indicated an intent to be executed, and because it was not
  effective even in principle when sourcing the script.

- Two contained `set -eu` commands affecting the behavior of the
  running interpreter, which were redundant because the script for
  which they are helpers already runs such a command. If they are
  later sourced into other scripts (a use for which they are not
  intended and would probably require further modification), those
  other scripts might not have these `set` commands, but in that
  case those commands' behavior would probably not be wanted.

This also adds "Must be sourced into the main journey test"
comments at the top of .sh files to which they apply, mirroring
the presence of such comments in other such files. Besides
consistency, this has the advantage of making clear what all the
scripts that are meant only for sourcing are being used for.