gix-date: make `time` a private dependency of `gix-date`
This should make the swap from `time` to `jiff` easier.

This comment[1] indicates that it's okay for `time` to be a public
dependency, but since this patch series is about swapping `time` for
`jiff`, it seemed appropriate to take this step first. And in
particular, it was *almost* already the case that `time` was a private
dependency of `gix-date`. The only thing we really had to button up was
the exposure of `time`'s custom formatting description language.

Jiff doesn't support `time`'s custom formatting machinery and instead
uses a strftime/strptime like API. We could expose that instead, but
since nothing (other than a test) was actually utilizing `time`'s custom
formatting machinery external to `gix-date`, I figured we might as well
completely encapsulate it.

[1]: https://github.com/Byron/gitoxide/issues/471#issuecomment-1328310133