Don't use `--` argument to `strip`
Unlike many utilities, the `strip` utility is not required by
POSIX to follow the XBD 12.2 Utility Syntax Guidelines, so need not
support `--`. On macOS, and possibly some other systems, it does
not, and `--` causes an error. Fortunately, it's okay to just not
pass it, because the value of `$TARGET_DIR` is known not to start
with `-`.