Create releases as drafts initially
This is inspired by the ripgrep workflow, though it currently does
it in a different way, since right now we are using actions rather
than the `gh` command.

Actual releases should of course become public. This can be done
manually; or another job, which depends on others, could be added
so it happens automatically if and after all builds succeed; or
this change to creating the releases as drafts (which at this
moment is mainly useful to facilitate testing without causing
confusion) could be reverted if no longer needed.