Allow the CI fuzz workflow to be manually triggered too
This adds workflow_dispatch as an event trigger for `cifuzz.yml`.

Since that should be sufficient to cover the rare case where one
wishes to run it outside of a pull request against the `main`
branch, this narrows the existing pull_request trigger to occur
only on `main`, as is already the case in the other workflows that
are intended to run more often than this one.