Put Ubuntu runner back to rolling for MSRV workflow
As of this commit, `ubuntu-latest` is `ubuntu-22.04`, with the
`ubuntu-24.04` runner image having beta status (even though Ubuntu
24.04 LTS is released), and `windows-latest` is `windows-2022`.

So this is putting Ubuntu back to 22.04, but it will automatically
become 24.04 evenutally, which the previous commit checks is likely
to work (since it would work now).

As detailed in #1514, there have not been problems in the past with
new Ubuntu versions on GHA runners breaking this workflow. In
contrast, there have been such problems with the Windows runners,
so this keeps the pin adjusted to the newer version, but it does
not change it to rolling.