Fix assertion message for joining `c:` to `relative`
This rewrites the custom message in that assertion so it states the
applicable semantics and avoids claiming that `c:relative` and
`c:\relative` are equivalent, which is not generally, nor usually,
the case.

(The message was also written with an unescaped `\r` in a non-raw
string literal, but a literal `\r` was intended. The new wording
no longer includes an explicit path, but if that is added then the
problem can be avoided by writing `\\r` or writing a raw string.)

This relates to:

- https://github.com/Byron/gitoxide/pull/1374/files/e955770c0b5c06a6c8518a06df4aa0cc3b506f16#r1721045107