Improve messages about joining `C:` on the right
Although the path `C:` is relative because it indicates the current
directory on the C: drive, it still "takes over" when concatenated
after at most if not all paths (including absolute paths).

When the paths it is concatenated with are absolute paths with
drive letters other than C:, this behavior, while not obvious, is
conceptually clear such that no other behavior could be correct.
This is because the (drive-specific) current directory on the C:
drive is not located on another drive, and therefore not found
under any path that specifies a drive that is not C:.

(There are other situations where `C:` on the right takes
precedence where the explanation is less elegant, but they are
mostly not relevant to the assertions currently present here.)

This relates to:

- https://github.com/Byron/gitoxide/pull/1374/files/e955770c0b5c06a6c8518a06df4aa0cc3b506f16#r1721055278
- https://github.com/Byron/gitoxide/pull/1374/files/e955770c0b5c06a6c8518a06df4aa0cc3b506f16#r1721079186