Clarify assertion messages involving `\\localhost`
This abbreviates the significance of `\\localhost` more
specifically, as win-absolute-unc-host rather than win-absolue-unc.

It also points out what seems unusual about joining `\` and
`\\localhost` to get `\localhost` with just one backslash. This
message should be further revised to state why this behavior
happens or otherwise describe it clearly, once known. (Or if this
turns out to be a bug in the standard library, then that can be
stated, with an issue link either in the code or commit message.)

This relates to:

- https://github.com/Byron/gitoxide/pull/1374/files/e955770c0b5c06a6c8518a06df4aa0cc3b506f16#r1721079809