fix!: remove all workspace dependencies
The problem is that with them, we don't notice anymore if the crate changes,
because a dependency changes. That also means that older versions of the dependency
may stay even though some other crates might pick up a newer version.

Ultimately, this will lead to drift and subtle incompatibilities.

We declare this breaking to enforce a proper re-release.