Make EXE_INFO testable and add a basic test for it
This extracts the code used to lazily initialize `EXE_INFO` into a
separate `exe_info` helper, which is an implementation detail, but
can be called any number of times without caching from the tests.

It also adds one test.

The tests of downstream functions (those that use `EXE_INFO`)
remain more important. But this will allow the test to be expanded,
or another added, to test more expectations that `EXE_INFO` is
supposed to satisfy.