Test EXE_INFO no local config even if temp dir is a repo
This adds a test that `EXE_INFO` does not return the path to a
local scope configuration file even if no other configuration file
is available and even if the temp dir (from `env::temp_dir()`) is
itself a Git repository.

This test is not quite done, because it is strangely passing, even
though the implementation is not yet hardened to the degree that it
should be able to avoid doing this. Specifically, although Git will
refuse to use (and, in `git config -l`, will omit) the local scope
configuration from a repository that is owned by another user (and
not allowed via `safe.directory`), this test is attempting to cause
a local repository owned by the current user to be used as the
directory `env::temp_dir()` will return.

Once that test bug is fixed, the test should start failing. Then,
when further hardening against unusual temporary directories (and
vulnerable `git` versions) is implemented, it should pass again.